Renaming of order classes
